% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lp_constraint.R
\name{add_lp_constraint}
\alias{add_lp_constraint}
\title{create LP constraints}
\usage{
add_lp_constraint(lp_obj)
}
\arguments{
\item{lp_obj}{the lp object contains test_inf and xlsx file for constraints}
}
\value{
a LP object
}
\description{
the objectives include "PASSAGE_SLICE", "FILTER", "CONTENT", "TIF", "TCC",
"MIN_OVERLAP", "ITEM_USAGE"
}
