% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addon_lp_constraint.R
\name{addon_accumulative_max_usage}
\alias{addon_accumulative_max_usage}
\title{add special LP constraints to an existing LP object}
\usage{
addon_accumulative_max_usage(lp_obj, max_usage_count)
}
\arguments{
\item{lp_obj}{the LP object that has been solved at least one time}
}
\value{
a LP object
}
\description{
add special LP constraints to an existing LP object
}
