% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addon_lp_constraint.R
\name{addon_blockSolution}
\alias{addon_blockSolution}
\title{add special LP constraints to an existing LP object--pair overlap control}
\usage{
addon_blockSolution(
  lp_obj,
  whichSolution = NULL,
  block_within_forms = T,
  block_across_forms = F,
  maxOverlapRate = 0.8,
  assignSolutionWtZero = T
)
}
\arguments{
\item{lp_obj}{the LP object that has been solved at least one time}

\item{whichSolution}{whichSolutiion in the lp_obj$formList should be added as a constraint controlling the overlap
rate between this solution and those being generated iteratively.}

\item{block_within_forms}{if TRUE, the overlap rate control per form is added as new constraints.}

\item{block_across_forms}{if TRUE, the overlap rate control across forms is added as a new constraint.
block_within_forms and block_across_forms can both be TRUE.}

\item{maxOverlapRate}{the maximum overlap rate between the next new solution and the previous solutions.}

\item{assignSolutionWtZero}{assign the solution items to have objective coef to be zeros.}
}
\value{
a LP object
}
\description{
add special LP constraints to an existing LP object--pair overlap control
}
