% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alias.R
\name{aliasListToTibble}
\alias{aliasListToTibble}
\title{convert the alias variable in the list format to a tibble format (original format)
that can be shown by a editDT}
\usage{
aliasListToTibble(alias_list)
}
\arguments{
\item{alias_list}{the alias list that has a variable "others" for all fields not initially needed
by the ATA}
}
\value{
the alias in tibble format
}
\description{
convert the alias variable in the list format to a tibble format (original format)
that can be shown by a editDT
}
