% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_template.R
\name{apply_template}
\alias{apply_template}
\title{Build the constraints using templates, which might include tcc, tif, bp, etc}
\usage{
apply_template(lp_obj)
}
\arguments{
\item{lp_obj}{the lp object contains the tcc file}
}
\value{
a LP object
}
\description{
Build the constraints using templates, which might include tcc, tif, bp, etc
}
