% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplex.R
\name{cplex}
\alias{cplex}
\title{Solve a linear program using the cplex solver}
\usage{
cplex(lp, timeout = 60, gap = 0, verbose = FALSE, solverCmd = NULL, ...)
}
\arguments{
\item{lp}{An R object that encapsulates the data used to specify a
linear program}

\item{timeout}{The number of CPU seconds before the solver must timeout}

\item{gap}{Absolute MIP gap}

\item{verbose}{Verbosity level}
}
\value{
Updated lp object
}
\description{
Solve a linear program using the cplex solver
}
\examples{
  setwd("/Users/ychien/Documents/code/R/MST/myFormAssembler")
  lp = readRDS("inst/rdsdata/Ex02.rds")
  timeout = 60
  gap = 0
  verbose = FALSE
  solverCmd = "cplex"
  result = cplex(lp, timeout, gap, verbose, solverCmd)

}
