% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst_utility.R
\name{createMSTResearchRDS}
\alias{createMSTResearchRDS}
\title{createMSTResearchRDS is to take items, content, alias, tic, tcc from an lp_obj and save to a rds
file that the CATSuite can read.}
\usage{
createMSTResearchRDS(lp_obj, rds_name)
}
\arguments{
\item{rds_name}{the rds name to be saved}

\item{lpObj}{the LP object}
}
\description{
createMSTResearchRDS is to take items, content, alias, tic, tcc from an lp_obj and save to a rds
file that the CATSuite can read.
}
