% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_panel.R
\name{do_panel}
\alias{do_panel}
\title{duplicate $constriant for panel and the $form_inf is modified accordingly
note that the $constraint$content must contain only one panel information in order to 
duplicate}
\usage{
do_panel(lp_obj)
}
\arguments{
\item{lpObj}{the LP object}
}
\value{
The updated LP object
}
\description{
duplicate $constriant for panel and the $form_inf is modified accordingly
note that the $constraint$content must contain only one panel information in order to 
duplicate
}
