% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{duplicated2}
\alias{duplicated2}
\title{A duplicated function that identifies if an element is unique or not}
\usage{
duplicated2(vec)
}
\arguments{
\item{vec}{a vector}
}
\value{
A vector of F/T
}
\description{
A duplicated function that identifies if an element is unique or not
}
