% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_fun.R
\name{expected_score}
\alias{expected_score}
\title{Calculate expected item scores for 3PL or GPC for multiple items
and a vector of theta values}
\usage{
expected_score(
  theta,
  IRT.par,
  IRT.model = "3PL",
  IRT.scale = 1.7,
  is_percent_scale = F
)
}
\arguments{
\item{theta}{A vector of ability values}

\item{IRT.par}{A matrix of IRT parameters. This should have the columns required by the IRT model of choice.}

\item{IRT.model}{A vector of character strings that specify the IRT model for each item.
This can also be a single value, which will be recycled.}

\item{IRT.scale}{A vector of scaling parameters used in the IRT models.
This can also be a single value, which will be recycled.}

\item{is_percent_scale}{if TRUE, the expected percent score for the given thetas will be returned, which is the tcc.}
}
\value{
An item by theta matrix of expected item scores
}
\description{
Calculate expected item scores for 3PL or GPC for multiple items
and a vector of theta values
}
