% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_fun.R
\name{expected_score_3PL}
\alias{expected_score_3PL}
\title{Calculate expected item scores for 3PL for a single item and
a vector of theta values}
\usage{
expected_score_3PL(theta, a, b, c, D = 1)
}
\description{
Calculate expected item scores for 3PL for a single item and
a vector of theta values
}
