% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_fun.R
\name{expected_score_GPC}
\alias{expected_score_GPC}
\title{Calculate expected item scores for GPC for a single item and
a vector of theta values}
\usage{
expected_score_GPC(theta, a, b, d, D = 1)
}
\description{
Calculate expected item scores for GPC for a single item and
a vector of theta values
}
