% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_fun.R
\name{fisher.3PL}
\alias{fisher.3PL}
\title{Calculate Fisher information for 3PL for a single item and
a vector of theta values}
\usage{
fisher.3PL(theta, a, b, c, D = 1)
}
\description{
Calculate Fisher information for 3PL for a single item and
a vector of theta values
}
