% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_fun.R
\name{fisher.GPC}
\alias{fisher.GPC}
\title{Calculate Fisher information for GPC for a single item and a vector
of theta values
The Calculation is based on Equation 10 in the Muraki (1993) paper,
which says that
      I(theta) = (a * D)^2 * var( score | theta )}
\usage{
fisher.GPC(theta, a, b, d, D = 1)
}
\description{
Calculate Fisher information for GPC for a single item and a vector
of theta values
The Calculation is based on Equation 10 in the Muraki (1993) paper,
which says that
      I(theta) = (a * D)^2 * var( score | theta )
}
