% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_fun.R
\name{fisher}
\alias{fisher}
\title{Calculate Fisher information for 3PL or GPC for multiple items
and a vector of theta values}
\usage{
fisher(theta, IRT.par, IRT.model = "3PL", IRT.scale = 1.7, getTestInf = F)
}
\arguments{
\item{theta}{A vector of ability values}

\item{IRT.par}{A matrix of IRT parameters. This should have the columns required by the
IRT model of choice.}

\item{IRT.model}{A vector of character strings that specify the IRT model for each item.
This can also be a single value, which will be recycled.}

\item{IRT.scale}{A vector of scaling parameters used in the IRT models.
This can also be a single value, which will be recycled.}

\item{getTestInf}{A boolean indicating if item-level information or test information is wanted.}
}
\value{
If getTestInf = F, an item by theta matrix of Fisher information values is return;
otherwise, a vector of test information vector is returned.
}
\description{
Calculate Fisher information for 3PL or GPC for multiple items
and a vector of theta values
}
