% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_constraint_across_forms.R
\name{gen_constraint_across_forms}
\alias{gen_constraint_across_forms}
\title{Constraint across forms}
\usage{
gen_constraint_across_forms(
  forms,
  n.items,
  x,
  x.coef,
  y = NULL,
  y.coef = NULL,
  type,
  rhs,
  label = "",
  consId = NULL,
  ys = NULL,
  ys.coef = NULL
)
}
\value{
A list of constraints
}
\description{
Constraint across forms
}
