% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_constraint_within_forms.R
\name{gen_constraint_within_forms}
\alias{gen_constraint_within_forms}
\title{Content Constraint within forms}
\usage{
gen_constraint_within_forms(
  forms,
  n.items,
  x,
  x.coef,
  type,
  rhs,
  label = "",
  consId = NULL,
  y = NULL,
  y.coef = NULL,
  ys = NULL,
  ys.coef = NULL
)
}
\value{
a list of constraints for each form
}
\description{
Content Constraint within forms
}
