% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_filter_constraint.R
\name{gen_filter_constraint}
\alias{gen_filter_constraint}
\title{Build the constraint data frame for a linear program from a filter vector.}
\usage{
gen_filter_constraint(x, label = "", consId = NULL)
}
\arguments{
\item{x}{A numeric vector with its length equal to the number of decision variables.
Each entry is either a numeric value or NA.}
}
\value{
A constraint data frame for the linear program
}
\description{
Build the constraint data frame for a linear program from a filter vector.
}
\details{
If x[i] is a numeric value then the i-th decision variable is fixed to be this value;
if x[i] is NA then no requirement is placed on the corresponding decision variable.
}
