% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_constraint_within_forms.R
\name{gen_psgcount_constraint_within_forms}
\alias{gen_psgcount_constraint_within_forms}
\title{Passage count constraint within forms}
\usage{
gen_psgcount_constraint_within_forms(
  forms,
  n.psgs,
  ys.start.at,
  ys,
  ys.coef,
  type,
  rhs,
  label = "",
  consId = NULL
)
}
\value{
a list of constraints for each form
}
\description{
Passage count constraint within forms
}
