% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst_utility.R
\name{generate_mst_all_path}
\alias{generate_mst_all_path}
\title{generate_mst_all_path is to generate all possible MST paths for a given MST design stage design
file that the CATSuite can read.}
\usage{
generate_mst_all_path(design = "")
}
\arguments{
\item{design}{a string with "-" as seperater. e.g. 2-3-4}
}
\value{
a data.frame with Stages and Paths
}
\description{
generate_mst_all_path is to generate all possible MST paths for a given MST design stage design
file that the CATSuite can read.
}
