% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_constraint_summary.R
\name{get_constraint_summary}
\alias{get_constraint_summary}
\title{Analysis constraints in the form}
\usage{
get_constraint_summary(lp_obj)
}
\arguments{
\item{lp_obj}{the lp object}
}
\description{
Analysis constraints in the form
}
