% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{get_def_alias}
\alias{get_def_alias}
\title{Get a list of default column aliases}
\usage{
get_def_alias(max_pts = 6)
}
\arguments{
\item{max_pts}{the number of irt_par_d columns that should be included}
}
\value{
list of default column aliases
}
\description{
Get a list of default column aliases
}
