% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_enemy.R
\name{get_ENEMY_COL_to_constraint}
\alias{get_ENEMY_COL_to_constraint}
\title{convert _ENEMY_COL to many constraints using the item column "ENEMY"}
\usage{
get_ENEMY_COL_to_constraint(lp_obj)
}
\arguments{
\item{lp_obj}{the lp object}
}
\value{
a LP object
}
\description{
convert _ENEMY_COL to many constraints using the item column "ENEMY"
}
