% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input.R
\name{get_input}
\alias{get_input}
\title{get input from R objects}
\usage{
get_input(lp_obj, input_tibble, add_default_alias = FALSE, verbose = F)
}
\arguments{
\item{input_tibble}{the input tibbles}

\item{add_default_alias}{if true, the default alias is used}

\item{lpObj}{the LP object}
}
\value{
The updated LP object
}
\description{
get input from R objects
}
