% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst_utility.R
\name{get_mst_path_from_content}
\alias{get_mst_path_from_content}
\title{get mst path from the content table}
\usage{
get_mst_path_from_content(contentTbl, n_module_in_a_test = 0)
}
\arguments{
\item{contentTbl}{the content table}

\item{n_module_in_a_test}{number of modules to form a MST test}
}
\value{
The updated LP object
}
\description{
get mst path from the content table
}
