% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glpk.R
\name{glpk}
\alias{glpk}
\title{Solve a linear program using GLPK.}
\usage{
glpk(lp, timeout = 60, gap.ratio = 0, verbose = FALSE, ...)
}
\arguments{
\item{lp}{An R object that encapsulates the data used to specify a
linear program}

\item{timeout}{The number of seconds before the solver must timeout}

\item{gap.ratio}{relative MIP gap tolerance}

\item{verbose}{Verbosity level}
}
\value{
Updated lp object
}
\description{
Solve a linear program using GLPK.
}
