% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_lp_obj.R
\name{init_lp_obj}
\alias{init_lp_obj}
\title{Initialize a lpObj object}
\usage{
init_lp_obj(
  basic,
  input = list(files = NULL, input_obj = NULL),
  objective,
  template = NULL,
  use_template = NULL,
  mst = NULL
)
}
\description{
Initialize a lpObj object
}
