% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example.R
\name{load_example}
\alias{load_example}
\title{Load an example (without solving it)}
\usage{
load_example(
  example_name = NULL,
  timeout = 60,
  solver = "cbc",
  gap = 0,
  verbose = TRUE
)
}
\description{
Load an example (without solving it)
}
