% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsolve.R
\name{lpsolve}
\alias{lpsolve}
\title{Solve a linear program using lpSolve.}
\usage{
lpsolve(
  lp,
  timeout = 10,
  presolve = c("rows", "cols"),
  gap = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{lp}{an R object that encapsulates the data used to specify a linear program model.
This is a list with the following components: constraint, objective, var.type,
upper and lower. The last three components are optional and default to NULL.}

\item{timeout}{either a positive integer value specifying the number of seconds after
which a timeout will occur, or zero, then no timeout will occur.}

\item{presolve}{a character vector specifying presolve steps to be carried out before solving}

\item{verbose}{either a character string or a TRUE/FALSE value to control the level
of reporting. The character string can be:
    "critical" (hard errors only),
    "severe" (errors only),
    "important" (warnings and errors),
    "normal",
    "detailed", or
    "full".
    TRUE means "normal", FALSE means "neutral".}

\item{...}{more control parameters to be passed to lpSolve}
}
\value{
Updated lp object
}
\description{
Solve a linear program using lpSolve.
}
