% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{merge2}
\alias{merge2}
\title{A merge function}
\usage{
merge2(x, y, all.x = FALSE)
}
\arguments{
\item{x}{the first data frame}

\item{y}{the second data frame}
}
\value{
A data frame with the same columns as x
}
\description{
1. An NA value in the first data frame matches to everything in the second data frame;
2. The ordering of both rows and columns in the first data frame is preserved.
}
