% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addon_lp_constraint.R
\name{modify_objective_for_Overall_max_IE_rate}
\alias{modify_objective_for_Overall_max_IE_rate}
\title{Modify objective coefficients to control the maximum item exposure rate in a series of solves}
\usage{
modify_objective_for_Overall_max_IE_rate(lp_obj)
}
\arguments{
\item{lp_obj}{the LP object that has been initialized and is ready to be solved.}
}
\value{
a LP object
}
\description{
Two ways of calculating the item exposure rates: one is based on the number of previous solves
(i.e. panels being built so far), and another way is use the total number of solutions that
the series of solves will achieve.
Two parameters are needed--lp_obj$objective_parameter$max_IE_rate and
lp_obj$objective_parameter$IE_rate_denominator (optional). If IE_rate_denominator = 0 or not
provided, the number of pervious solves will be used for calculating the item exposure rates.
}
