% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mutate_col_type}
\alias{mutate_col_type}
\title{change column data types}
\usage{
mutate_col_type(tbl1, col_types)
}
\arguments{
\item{tbl1}{a regular data frame or a tibble}

\item{col_types}{the column types include 'c' for character, 'd' for double,
'i' for integer, 'l' for logical, and 'f' for factor.}
}
\value{
a tibble or a data frame with changed column types
}
\description{
change column data types
}
