% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_lp_constraint.R
\name{mutate_lp_constraint}
\alias{mutate_lp_constraint}
\title{mutate LP constraints}
\usage{
mutate_lp_constraint(lp_obj)
}
\arguments{
\item{lp_obj}{the lp object contains test_inf and xlsx file for constraints}
}
\value{
a LP object
}
\description{
$constraint$mutator may operate on constraints with regular ATTRIBUTE or _CLUSTER.
This feature depends on the user to provide a careful designed mutator spec; otherwise, it may
cause no feasible solution.
}
