% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alias.R
\name{read_alias}
\alias{read_alias}
\title{Read alias from a file or from an tibble variable}
\usage{
read_alias(lp_obj, input_obj = NULL)
}
\arguments{
\item{lp_obj}{the lp object}

\item{input_obj}{the alias object}
}
\value{
a lp object
}
\description{
Read alias from a file or from an tibble variable
}
