% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_constraint.R
\name{read_constraint}
\alias{read_constraint}
\title{Read the constraints from a xlsx/xls file or a google sheet file}
\usage{
read_constraint(lp_obj, input_obj = NULL)
}
\arguments{
\item{lp_obj}{the lp object contains testInf and xlsx file for constraints}

\item{input_obj}{the constraint object}
}
\value{
a LP object
}
\description{
Read the constraints from a xlsx/xls file or a google sheet file
}
