% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_enemy.R
\name{read_enemy}
\alias{read_enemy}
\title{Read the enemy groups from a xlsx/xls file or a google sheet file}
\usage{
read_enemy(lp_obj, input_obj = NULL)
}
\arguments{
\item{lp_obj}{the lp object contains testInf and xlsx file for enemy groups}

\item{input_obj}{the enemy groups object}
}
\value{
a LP object
}
\description{
Read the enemy groups from a xlsx/xls file or a google sheet file
}
