% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gs.R
\name{read_gs}
\alias{read_gs}
\title{Read in google sheets from a given title}
\usage{
read_gs(gs_file, sh, ...)
}
\arguments{
\item{gs_file}{the gsheet file name}

\item{sh}{a vector of positive integer or character string specifying index or title.
use sh == 0 to get all sheets}
}
\value{
a list of tibbles for the sheets requested
}
\description{
Read in google sheets from a given title
}
