% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_input}
\alias{read_input}
\title{read in xlsx/xls files or google sheet files}
\usage{
read_input(lp_obj, add_default_alias = FALSE)
}
\arguments{
\item{lp_obj}{lp_obj$files might contain alias_file, pool_file, cons_file, slice_file,
passage_slice_file, tcc_file, tif_file}
}
\value{
The updated LP object
}
\description{
read in xlsx/xls files or google sheet files
}
