% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mutator.R
\name{read_mutator}
\alias{read_mutator}
\title{Read the mutators from a xlsx/xls file or a google sheet file}
\usage{
read_mutator(lp_obj, input_obj = NULL)
}
\arguments{
\item{lp_obj}{the lp object contains xlsx file for mutators}

\item{input_obj}{the mutators object}
}
\value{
a LP object
}
\description{
Read the mutators from a xlsx/xls file or a google sheet file
}
