% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_passage_slice.R
\name{read_passage_slice}
\alias{read_passage_slice}
\title{Read the passage slice from a xlsx/xls file or a google sheet file}
\usage{
read_passage_slice(lp_obj, input_obj = NULL)
}
\arguments{
\item{lp_obj}{the lp object contains testInf and xlsx file for slice}

\item{input_obj}{the slice object}
}
\value{
a LP object
}
\description{
Read the passage slice from a xlsx/xls file or a google sheet file
}
\examples{
  lp_obj = init_lp_obj(
    basic = list(
      test_name = "read_passage_slice example",
      n_form = 1),
    input = list(files = list(
      passage_slice_file = list(file = "inst/extdata/Psg01/passage_slice.xlsx")
    )),
    objective = c("passage_slice")
  )
}
