% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pool.R
\name{read_pool}
\alias{read_pool}
\title{Read items from a xlsx/xls file or a google sheet file}
\usage{
read_pool(lp_obj, input_obj = NULL)
}
\arguments{
\item{lp_obj}{the lp object contains testInf and xlsx file for items}

\item{input_obj}{the item/pool object}
}
\value{
a lp object
}
\description{
Read items from a xlsx/xls file or a google sheet file
}
