% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slice.R
\name{read_slice}
\alias{read_slice}
\title{Read the slice from a xlsx/xls file or a google sheet file}
\usage{
read_slice(lp_obj, input_obj = NULL, verbose = F)
}
\arguments{
\item{lp_obj}{the lp object contains testInf and xlsx file for slice}

\item{input_obj}{the slice object}
}
\value{
a LP object
}
\description{
Read the slice from a xlsx/xls file or a google sheet file
}
