% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tcc.R
\name{read_tcc}
\alias{read_tcc}
\title{Read the tcc target from a xlsx/xls file or a google sheet file}
\usage{
read_tcc(lp_obj, input_obj = NULL, verbose = FALSE)
}
\arguments{
\item{lp_obj}{the lp object contains the tcc file}

\item{input_obj}{the tcc object}
}
\value{
a LP object
}
\description{
Read the tcc target from a xlsx/xls file or a google sheet file
}
