% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tif.R
\name{read_tif}
\alias{read_tif}
\title{Read the tif target from a xlsx/xls file or a google sheet file}
\usage{
read_tif(lp_obj, input_obj = NULL, verbose = F)
}
\arguments{
\item{lp_obj}{the lp object contains the tif file}

\item{input_obj}{the tif object}
}
\value{
a LP object
}
\description{
Read the tif target from a xlsx/xls file or a google sheet file
}
