% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_glpk.R
\name{set_glpk}
\alias{set_glpk}
\title{Set GLPK linear program model object}
\usage{
set_glpk(lp)
}
\arguments{
\item{lp}{An R object that encapsulates the data used to specify a linear program model.
This is a list with the following components:
constraint, objective, var.type, upper and lower. The last three components
are optional and default to NULL.}
}
\value{
A glpkAPI::glpkPtr object.
}
\description{
Set up a new GLPK linear program model object from a corresponding
R object.
}
