% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lpsolve.R
\name{set_lpsolve}
\alias{set_lpsolve}
\title{Set lpSolve linear program model object}
\usage{
set_lpsolve(lp)
}
\arguments{
\item{lp}{An R object that encapsulates the data used to specify a linear program model.
This is a list with the following components: constraint, objective, var.type,
upper and lower. The last three components are optional and default to NULL.}
}
\value{
A lpSolveAPI::lpExtPtr object. This is an external pointer to a "lprec" structure
used by the lp_solve C API that underlies the lpSolveAPI package.
}
\description{
Set up a new lpSolve linear program model object from a corresponding
R object.
}
