% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_lp}
\alias{solve_lp}
\title{solve an initialized LP object}
\usage{
solve_lp(lp_obj, timeout = NULL, solver = NULL, gap = NULL, verbose = T, ...)
}
\arguments{
\item{lp_obj}{the LP object}

\item{timeout}{default is 60 seconds}

\item{solver}{"cbc" (default), "symphony", "glpk" and "lpsolve".}

\item{gap}{absolute MIP gap with default = 0}

\item{verbose}{}
}
\value{
The updated LP object with an "x" component, which is
a binary indicator matrix that defines the items picked for forms.
}
\description{
solve an initialized LP object
}
