% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_example.R
\name{to_solve}
\alias{to_solve}
\title{to solve an initialized lp object}
\usage{
to_solve(lp_obj, timeout = NULL, solver = NULL, gap = NULL, verbose = NULL)
}
\arguments{
\item{timeout}{default is 60 seconds}

\item{solver}{"cbc" (default), "symphony", "glpk" and "lpsolve".}

\item{gap}{absolute MIP gap with default = 0}

\item{verbose}{}

\item{lpObj}{the LP object}
}
\value{
the solved lp object
}
\description{
to solve an initialized lp object
}
