% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_cont.R
\name{to_solve_cont}
\alias{to_solve_cont}
\title{to solve an initialized lp object iteratively}
\usage{
to_solve_cont(
  lp_obj,
  block_within_forms = T,
  block_across_forms = T,
  maxOverlapRate = 0.8,
  max_usage = 1,
  timeout = NULL,
  solver = NULL,
  gap = NULL,
  maxIter = 10,
  exist_n_iter = 0,
  verbose = T,
  ...
)
}
\arguments{
\item{block_within_forms}{default is T.}

\item{block_across_forms}{default is T.}

\item{maxOverlapRate}{the maximum overlap rate between the new solution and the blocked solution.
The maximum number of overlapping items is a rounded integer of maxOverlapRate * number of items per form.}

\item{max_usage}{the maximum item usage in maxIter}

\item{timeout}{default is 60 seconds}

\item{solver}{"cbc" (default), "symphony", "glpk" and "lpsolve".}

\item{gap}{absolute MIP gap with default = 0}

\item{maxIter}{maximum number of iterations}

\item{exist_n_iter}{the number of existing iterations}

\item{verbose}{}

\item{lpObj}{the LP object that has been initialized and is ready to be solved.}
}
\value{
the solved lp object; if forms are successfully built, they can be found from formList,
consSummaryList, and formSummaryList under lp_obj.
}
\description{
to solve an initialized lp object iteratively
}
\examples{
  example_name = "MST08_tiny"
  solver = "cbc"
  timeout = 60
  gap = 0
  verbose = TRUE
  maxIter = 4
  maxOverlapRate = .50
  lp_obj2 = myFormAssembler::run_example(example_name, T, timeout, solver, gap, verbose)
  lp_obj2 = to_solve_cont(lp_obj2, block_within_forms = T, block_across_forms = T, maxOverlapRate = maxOverlapRate, timeout = timeout, solver = solver, maxIter = maxIter, gap = gap)
}
